Create the flatmap package
