+++ flatmap/doc.go
@@ -1,0 +1,11 @@
+/*
+Package flatmap contains functions to flatten and unflatten maps:
+
+  - [Do] flattens a nested map into a flat map.
+  - [Undo] unflattens a flat map into a nested map.
+
+The flattening roughly follows the [JSONPath] standard.
+
+[JSONPath]: https://goessner.net/articles/JsonPath/
+*/
+package flatmap
